basic display and movement from players

+++ examples/19 - snake with guns/source/audiomanager.d
@@ -173,7 +173,7 @@ tron.AudioManager = function()
 		this._nFocus = 0;		
 	}
 	
-	void addFocus(int x, int y, int dist)
+	void addFocus(int x, int y, double dist)
 	{
 		/*
 		var i = this._nFocus++;
